move all tests into unit
